Update vgg and resnet via api v2
